Library-dependent features are converted to GS geometric elements before the matching phase
Before that, the conversion happened after a feature has been accepted (the main idea was to avoid converting useless data).
This structure required the user to manipulate different data formats when specifying the matching function and, as such, to
be aware of the different libraries capabilities (XOM, Geotools). By converting to a standard geometric format earlier in the
pipeline, all of this is hidden from the user and he only manipulates Points and Lines.